requirements.txt
Django==4.2.11
djangorestframework==3.14.0 
mysqlclient==2.2.4
Pillow==10.3.0